refactor: Move constraints to different file
