+++ src/constraints/constrainted_type.rs
@@ -1,6 +1,6 @@
-use rand::{distr::Uniform, RngCore};
-
 use crate::{evaluator::{context::{Constraint, Context}, eval_error::EvalError, evaluator::evaluate_expression, object::Object}, parser::ast::{ConstraintKind, DataType, DataTypeKind}};
+
+use super::constraints::{MultipleOfConstraint, RangeConstraint};
 
 #[derive(Debug)]
 pub struct ConstrainedType {
@@ -36,62 +36,5 @@ impl ConstrainedType {
         }
 
         Ok(ConstrainedType { type_kind: data_type.kind, constraints: evaluated_constraints })
-    }
-    
-    pub fn generate(&self, _rng: &mut dyn RngCore, _max_attempts: usize) -> Option<Object> {
-        todo!()
-    }
-}
-
-
-#[derive(Debug, Clone)]
-pub struct RangeConstraint {
-    pub min: isize,
-    pub max: isize,
-}
-
-impl RangeConstraint {
-    fn new(min: isize, max: isize) -> Self {
-        RangeConstraint { min, max }
-    }
-}
-
-impl Constraint for RangeConstraint {
-    fn validate(&self, object: &Object) -> bool {
-        matches!(object, Object::Int(v) if *v >= self.min && *v <= self.max)
-    }
-
-    fn description(&self) -> String {
-        format!("range={}..={}", self.min, self.max)
-    }
-    
-    fn build_sampler(&self) -> Option<Box<dyn std::any::Any>> {
-        let dist = Uniform::new_inclusive(self.min as i32, self.max as i32);
-        Some(Box::new(dist))
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct MultipleOfConstraint {
-    pub base: i32,
-}
-
-impl MultipleOfConstraint {
-    pub fn new(base: i32) -> Self {
-        MultipleOfConstraint { base }
-    }
-}
-
-impl Constraint for MultipleOfConstraint {
-    fn validate(&self, value: &Object) -> bool {
-        matches!(value, Object::Int(v) if *v as i32 % self.base == 0)
-    }
-
-    fn description(&self) -> String {
-        todo!()
-    }
-
-    fn build_sampler(&self) -> Option<Box<dyn std::any::Any>> {
-        None
     }
 }
