+++ src/constraints/mod.rs
@@ -1,2 +1,3 @@
 pub mod constrainted_type;
+pub mod constraints;
 pub mod tests;
